refactor(liferay): move shared resource queries and paths into portal
