+++ src/features/liferay/resource/liferay-resource-get-template.ts
@@ -4,9 +4,9 @@ import type {HttpApiClient} from '../../../core/http/client.js';
 import {LiferayErrors} from '../errors/index.js';
 import {runLiferayInventoryTemplates} from '../inventory/liferay-inventory-templates.js';
 import {buildSiteChain} from '../portal/site-resolution.js';
+import type {DdmTemplatePayload} from '../portal/template-queries.js';
 import {resolveResourceSite, listDdmTemplates} from './liferay-resource-shared.js';
 import {matchesDdmTemplate, matchesInventoryTemplate} from '../liferay-identifiers.js';
-import type {DdmTemplatePayload} from './liferay-resource-payloads.js';
 
 type ResourceDependencies = {
   apiClient?: HttpApiClient;
